[raft] fix flaky RemoveDeadReplica test (#9301)
Occasionally when the driver runs, it didn't recognize the dead replica
because
of delay in gossip.
1. Sometimes, it will see the replica as suspect, in these case, I
change the driver to remove suspects when down replicating.
2. The driver previously can also get stuck if it removed a healthy
replica, and then when it tries to replace dead replica, it cannot find
a node. See
https://github.com/buildbuddy-io/buildbuddy-internal/issues/4960 I
change the driver code so that if we will finish the removal first for
AddReplica and ReplaceDeadReplica if there is replicas in the progress
of removal.

Also, I am skipping `TestCleanupZombieReplicaNotInRangeDescriptor` for
now. This test need rework b/c the current set up is not realistic with
the introduction of staging replica descriptors.

Success over 500 runs:

https://app.buildbuddy.io/invocation/483989ec-1c32-417e-b30a-126b728b3a23