[raft] store_test: remove out-of-range check from getReplica (#9302)
We will get out of range error when we call getReplica when we add a
replica and
try to get the replica between 1) the time when the config change is
committed at raft, the replica is created and the replica is caught up.
This can cause TestUpReplicate to fail even when the replica is added
successfully.

https://github.com/buildbuddy-io/buildbuddy-internal/issues/4192