Fix frontend JS/CSS caching and add tests (#9319)
We're serving stale JS/CSS to the UI because the `sha` rule is broken,
resulting in the `hash` URL parameter always having the same value
(`5cd33...`), instead of matching the actual content hash.

The `sha` rule was broken because the `find` command wasn't following
symlinks, so it was producing no files. As a result, the hash that we
are computing is `5cd33...` which is the same as what we get from
running `printf '' | shasum | shasum`

The actual fix in this PR is to just tack on the `-L` flag to `find`;
everything else is just testing.

Fixes https://github.com/buildbuddy-io/buildbuddy-internal/issues/4914