Fix a bug in migration_cache.doubleWriter
Since https://github.com/buildbuddy-io/buildbuddy/pull/9242, the destination writer `Write` method never fails, so destDeleteFn will never be called.

However, even before that PR, it was possible for the source Commit to fail while the destination Commit succeeds.

This PR makes it so that if any source write or commit fails, we delete the value from the destination.

TBH, I'm not sure why we need to delete from the destination in this case, but this brings the behavior closer to how it was before my optimizations.