Codesearch server only processes one update per repo at a time (#9327)
There are numerous race conditions that can occur when re-indexing the
same repo concurrently - the same files may get added twice in many
circumstances.

This change makes the codesearch server lock on a per-repo basis. On
it's own, this is not a scalable change, especially for large repos.
However, it is made in preparation for implementing incremental index
updates.