Fix `OrderedArgs.Append` (#9344)
`Append` was previously incorrectly appending command-options to before
the command, when it ought to have appended to before the double-dash,
if one exists.

Additionally, it is the job of the parser, not the classifier, to
identify the first double dash, if one exists, which simplifies
classification significantly and makes appending configs without
prematurely re-ordering args much, much easier.