Factor common CS doc update/delete functionality  (#9341)
This does change the `DeleteDocument` semantics slightly - instead of
being applied immediately the index, they are now added to the current
batch (consistent with `AddDocument` and `UpdateDocument`). We're not
currently calling it from anywhere outside of `index.go`, so this should
be safe.

Note that the `lookupDocId` helper function isn't attached to `Writer`
because it will also be used from `Reader` in an imminent follow on
change.