[cachetools] UploadFromReader always returns number of bytes (confirmed) uploaded (#9358)
The current state of `master`:
* When using zstd compression, `uploadFromReader` will report 0 bytes
uploaded if uploading a blob that already exists.
* When not using compression, `uploadFromReader` will report the number
of bytes sent to the server.

This change modifies `uploadFromReader` to _always_ report the number of
bytes sent to the server.
AFAICT, the only function that depends on this number is
`snaputil.Cache`. And the callers of `snaputil.Cache` use the reported
number of bytes uploaded for tracing and logging.

In addition, this change
* Adds test coverage for calling `UploadFromReader` when a blob already
exists in the CAS.
* Reverts `uploadWriteCloser`; will re-land after these tests are in.