build_event_handler: smarter persisted test artifacts (#9291)
These are artifacts created by TestRunner action defined by Bazel's
native Java
code here (1). It's usually a fixed list of optional outputs, so it
should be
relatively safe to bump the limit by 10x (we will do that in a separate
PR).

In case the limit is met, prioritize saving outputs from failed tests
over outputs
from passed tests. We differentiate the twos by checking if the test
status is
PASSED or not.

(1):
https://cs.opensource.google/bazel/bazel/+/615b63dd005923e12e1356c1368e80dd121f2c5a:src/main/java/com/google/devtools/build/lib/analysis/test/TestRunnerAction.java;l=402