CLI: Error out when attempting to expand unrecognized config option (#9355)
It is clearer to fail early, and means we do not need to create a
potentially confusing partially-expanded list of command-line arguments
to pass to bazel. The goal is that any command line that is actually
passed to bazel is fully expanded and free from any `rc` options, save
the single `ignore_all_rc` option we add when consuming the rc files.