Track and expose the most recently indexed commit per repository in codesearch (#9324)
This change adds the ability for index.Writer and index.Reader to record
and produce the most recently indexed commit, per repository.

It also exposes this via a `RepoStatus` rpc in the codeearch service.

This is in support of a near-future change which will enable incremental
updates of the codesearch index as individual commits are created on the
indexed branch.

The factoring here could be improved in the future - all the core
codesearch types (e.g. IndexReader/IndexWriter) should probably remain
completely ignorant of anything git-specific, such as commits.
Additionally, the logic for tracking commits is duplicated across the
CLI tool and the server, and doesn't really belong directly in either of
them. This should probably be addressed by introducing a more explicit
git-specific layer in between the clients (server, CLI) and the core
codesearch implementation.