Improve lockmap write benchmark
Previously, it was mostly benchmarking `random.String(64)`. I pulled that out of the loop by pre-creating keys. I also made the number of keys variable, so we can have benchmarks for high and low contention.

I didn't modify the read benchmark because I think it's good to have one benchmark that has 0 contention.