Codesearch: Add support for incremental indexing of git repos (#9372)
This change adds server-side support for incremental indexing to
codesearch.

The overall design here is:
- The CS server keeps track of the most recent commit it has indexed
- Clients can send in a list of commits (files added/updated, and names
of files deleted).
- The server will apply these changes, as long as the most recently
indexed commit is a parent of one of the commits sent
- Already-indexed commits are skipped

This design is a little funky, and that funkiness is driven by the fact
that CI jobs will have access to the git repo, but the codesearch server
will not. So the CI jobs will compute the set of files that need to be
updated for any given change, and the CS server will apply those
updates. It would be cleaner if the CS server itself could be asked to
incrementally update a branch/repo, but that would involve the CS server
downloading the entire repo (including git history - too much data to be
pulled every commit), or making a number of GH API calls (which has
quota concerns). So, here we are.