Allow reading in-progress executions from Redis (#9395)
Today, if we were to start storing in-progress execution metadata in
Redis instead of MySQL, the UI would not be able to view in-progress
executions. We do already read execution state from Redis, but only for
executions that have completed (and are sitting in Redis waiting for the
invocation to complete).

This PR makes it so we read the in-progress execution state from Redis
as well, so that in-progress executions can be displayed in the
executions page.