Add a GROUP_ACCESS capability.
Today users and API keys in a group have broad access to various APIs by
virtue of group membership. This PR makes this access more explicit by
modeling it as its own capability.

Users always have the GROUP_ACCESS capability and most API key types
have this capability as well, except the freshly added "audit log
reader" capability.

This PR introduces the `AuthorizeGroupAccess` function which is meant to
be a drop-in replacement for `authenticator.AuthenticatedUser` with the
addition of checking the GROUP_ACCESS capability.
The existing `AuthorizeGroupAccess` is renamed to `AuthorizeTargetGroupAccess`
as the existing function is not frequently used and in most cases can be
replaced by a call to the new function.