[Macs] revamp the mac monitoring dashboard (#9399)
This PR overhauls the mac monitoring dashboard. The key changes are that
it:
* Makes the `inventory_hostname_pattern` work in all graphs.
* Adds a new `group_name` variable for easy filtering by group name.
* Makes the four bottom rows (Node disk/network IO, Node memory usage,
Node swap usage, Node swap rate) repeat by `group_name` rather than
`inventory_hostname`. This significantly reduces the number of graphs
that need to be rendered when these rows are expanded, making them load
much faster.
* Removes the `inventory_hostname` variable.
* Adds graphs showing disk usage by group.
* Removes some graphs that don't work on mac.

This should make the dashboard a fair bit easier to use and more useful,
especially for oncall.