Codesearch incremental update extraction tools (#9403)
This change adds some functions that will be used to create incremental
update payloads for the codesearch indexer. This will eventually be
wired into the BB CLI, and will be run on a local clone of a git
repository.

These functions look at a range of commits on a single branch, and
extract the changes from each commit in the range that need to be sent
to the indexer: the contents of all added/modified files, and the names
of deleted / renamed files.