CLI: Refactor config parsing to use the new parsing constructs (#9389)
This change is a little odd, since many of the refactored functions in
`parser.go` now belong in `parsed.go`, but I elected to make the changes
first and then move them in a later PR so as to make viewing the diffs
easier when reviewing.

Primarily, this change:
- Directly uses the options-retrieval abilities of parsed.Args to
consume the rc file options
- Parses the rc files into `parsed.Config` structs in a map keyed by
config name and into a default config struct for unnamed configs (for
example, `common` or `build` as opposed to `common:foo` or `build:bar`)
- Uses these config structs to expand config options

Small behavior changes include that `common:` will no longer be
equivalent to `common`, which mirrors bazel's bahavior, and that the
`always` phase is now supported.

Once this is approved, a follow-on PR will move the functions indicated
with the relevant `TODO`s to their appropriate locations with minimal
changes as are necessary to move them as described.

---------

Co-authored-by: Brandon Duffany <brandon@buildbuddy.io>