[cachetools] add support for zstd to UploadWriter (#9436)
For writing OCI image layers to the CAS, I'll need zstd support, so I
went ahead and added it. This change also includes tests for using zstd
with `UploadWriter`.

`UploadWriter` keeps a `cbuf` for writing compressed bytes.
`CompressZstd` may allocate an additional buffer if `cbuf` is not large
enough for the compressed output. However, I do not want `UploadWriter`
to take ownership of that buffer: I want to keep a reference to the same
`cbuf` so I can return it to the pool on `Close`.