api_server: Use callback for invocation lookup (#9307)
Replace LookupInvocation with LookupInvocationWithCallback in
GetInvocation, GetTarget, and GetAction API endpoints. This avoids
loading the full invocation proto into memory, which can be large,
significantly reducing memory usage. The callback processes events
iteratively.

Specifically for GetTarget, we also move the TargetSelector filtering
logic inside the callback to reduce the size of the TargetMap when a
selector is available.

Change maxEventCount in LookupInvocation callback filter to a
configurable flag with the same default value. This enable us to test
the changes with a much lower maxEventCount to trigger the logic more
eagerly.