[oci cache] replace WriteBlobToCache with NewBlobUploader (#9438)
Instead of constructing a `TeeReader(upstream, response)` and passing it
to `UploadFromReader`, write the OCI blobs to the given writer
(currently, an HTTP response) and make a best effort to write them to
the cache.

If we encounter an error writing to the cache (or the CAS makes a short
read because the blob already exists), we still have a prayer of
returning the entire blob to the caller.