invocation_target_group_card: show cache status
Adds a badge next to the target label in the invocation targets
list to display test cache status information.

The badge shows "Cached" if all runs were served from cache.
It shows "X cached / Y runs" if some runs were cached, where X
is the number of cached runs and Y is the total number of runs.
This case is relatively rare as test shards, runs and attempts usually
share the same cache TTL.

Nothing is displayed if no runs were cached.

Note that we are not currently using test cache status to sort the
target group. Such logic can be added later in a future change. Also,
there is insufficient information from Build Event Stream to indicate
whether a build target was cached or not, so all build targets are being
shown without the badge, which could be misinterpreted as no-cache.