Only set metadata from execution summary if it's valid (#9447)
decodeMetadataFromExecutionSummary can return nil and overwrite existing
metadata

I noticed this bug from a unit test on a PR I'm working on, but the
comment about the execution summary is 3 years old. @tempoz can we just
remove this code? Or are we still concerned about older on-prem
customers who haven't upgraded?