Read execution data from Redis to flush to Clickhouse (#9391)
Currently when an executor has finished execution, it sends a payload to
the execution server which immediately flushes the data to clickhouse

In order to track runner recycling time in clickhouse, which may occur
after the requested execution has finished, this PR first stores
execution data in Redis. It then reads the data from Redis and flushes
it to Clickhouse. In itself this PR should functionally be a no-op.

Next, my plan is to 
1. Move flushing data to clickhouse to after the client closes the
PublishOperation stream
2. Executor publishes another update with runner recycling time. The
execution server updates the in-progress execution data in Redis. Now,
when the data is read in step 1, it will include the runner recycling
time