[oci resolver] separate AuthenticateWithRegistry from Resolve (#9471)
In `ociconv`, the `authenticateWithRegistry` function resolves an image
to check that credentials are valid, but does not use the returned
image. This change creates a separate `AuthenticateWithRegistry`
function that issues an HTTP HEAD request to check the credentials and
does not otherwise fetch any data from the server.

This change also adds info logs for `AuthenticateWithRegistry` and
`Resolve` indicating which upstream registry we're making requests to.