Integrate TreeFetcher with vfs.
The fetcher that's setup by the runner is passed to the vfs code instead
of using a separate download implementation in the vfs code.

Since the downloads are not guaranteed to by finished by the end of
DownloadInputs, the collection of transfer stats is delayed until after
Run is finished for both the vfs and non-vfs paths.