Add incremental indexing to codesearch-enabled workflows (#9470)
This PR adds a new action to workflows from codesearch-enabled
organizations. This new action uses the `bb index` command to compute
and send an incremental update to the codesearch server.

Note that we explicitly avoid recycling kythe action runners, but we
don't do the same for this new action, because this new action only
requires access to the checked-out git repo, it's fast, and it doesn't
generate any extra files.