Add several codesearch index utilities (#9488)
Adds 3 functions:
- `DropNamespace`, which drops an entire namespace
- `DeleteMatchingDocuments`, which deletes all docs that match a given
field. Intended to be used to delete an entire repository at once
- `CompactDeletes`, which removes deleted doc ids from all posting lists
in a namespace. Note that this is a potentially very expensive operation
for large indices, but it will keep the index size proportional to the
size of the repo, rather than the number of updates that have been
applied.

None of these are used or exposed anywhere yet.