Record network metrics for VMs and OCI containers (#9502)
Depends on https://github.com/buildbuddy-io/buildbuddy/pull/9500

Captures `{tx,rx}_{bytes,packets}` and populates the UsageStats proto
with this information (behind a flag).

Stats are collected from the host side of the veth pair to avoid having
to enter a netns to collect stats from the namespaced side, which has
too much overhead (especially since we probably want to poll these and
show a chart, similar to disk usage). This means that we have to swap
the tx/rx traffic for it to reflect the container's perspective, which
is a bit of a hack, but IMO an OK tradeoff for now.