cli/fix: avoid using langs.go twice
Replace #9472.

We observed that cli/test/integration/compact:compact_test failed
consistently in our 'bazel coverage' scheduled pipeline. There are a few
layers to this onion which we need to peel:

First, the error was that our testcli.CombinedOutput was printing an
extra line at the beginning of the stdout:

```
Diff:
--- Expected
+++ Actual
@@ -1 +1,2 @@
+Already covered cli/fix/langs/langs.go
{
```

This line came from @rules_go//go/tools/coverdata/coverdata.go
through the RegisterFile func. This suggests that our testcli was built
under coverage-enabled mode, which is definitely unintended (!).

However, that's not the problem we are fixing here. We are fixing the
fact that `cli/fix/langs/langs.go` is being used twice in our build: one
for the //cli/fix/langs:langs package and one for the
//cli/fix/langs:gazelle package. Achieve this by fully replacing :langs
with :gazelle and switch the BUILD file of cli/fix/langs to manually
managed.