Remove CloseAndRecv in runtime.Cleanup
This was added in https://github.com/buildbuddy-io/buildbuddy/pull/9503 because I thought it might fix our memory leak, but it wasn't related.

Also, it does prevent the active RPC count from continuously going up, but it causes it to continuously go down, though at a slower rate.

I suspect that the prometheus metrics interceptor is incrementing the `handled` count every time `CloseAndRecv` is called, so I just added a `CloseAndRecv` call in `Close`, if it hasn't been called already. Cancelling the stream context, should be sufficient to close the stream, but this way, `distributed_client.RemoteWriter` doesn't rely on that, and as long as `Close` is called, the stream will be closed.