Indexing from CLI now falls back to full reindex if necessary (#9517)
Notably, this PR adds a check for codesearch-enablement, and injection
of the repo access token, to the main app server, before forwarding the
indexing request to the codesearch server.

I was able to mostly test this locally, but I wasn't able to launch the
server with the github integration enabled, so I'm not confident in
codesearch.go `getGitRepoAccessToken` code. I pulled most of this from
[service.go](https://github.com/buildbuddy-io/buildbuddy/blob/master/enterprise/server/workflow/service/service.go#L629).