Fix 2 codesearch incremental bugs (#9549)
1. When we clone repos for workflows, we clone with `depth=1`, which
makes history unavailable. This change adds an `--unshallow` fetch to
the codesearch incremental update action, which fetches history (but
still avoids fetching blobs from history). Note we might be able to do
this entire thing slightly more cheaply with a treeless clone
(`--filter=tree:0`), as opposed to a blobless clone, but I'm sticking
with the current blobless clone that's used already for now.

2. The parsing of `git whatchanged` is made more robust. The git output
in the workflows included a warning about forced updates not being
shown, probably due to some bit of config from the workflow setup. This
output was breaking the parsing. We now parse the rows with regexes to
be more resilient. The next most resilient step here would probably be
trying to work programmatically with the git repo directly, rather than
through the command line client, but this would be a step up in
complexity, so we'll see how this does.