tweak deflake settings (#9525)
I found test_output=errors a bit inconvenient when working with tests
with a lot
of logs like raft tests. To investigate these errors, I ususally open
the test
log files with an editor. But the file is shown before all the log
lines. And
sometimes, when there are a lot of logs, I cannot get to the file.