Don't parse Bazel RequestMetadata thrice per server request (#9565)
In
[grpc_server.CommonGRPCServerOptionsWithConfig](https://github.com/buildbuddy-io/buildbuddy/blob/9e92b293bf56fbdc9584c2f65a88f4a8e19ee923/server/util/grpc_server/grpc_server.go#L245-L258),
`propagateRequestMetadataIDsToSpanUnaryServerInterceptor()` and its
streaming counterpart are added to the interceptor chain before
`copyHeadersUnaryServerInterceptor()` which is added in
`interceptors.GetUnaryInterceptor`/`interceptors.GetServerInterceptor`.
The `copyHeaders` interceptors call
`bazel_request.ParseRequestMetadataOnce(ctx)`. This means that the
`propagateRequestMetadataIDsToSpan` interceptors don't benefit from the
fact that `bazel_request` stores the parsed Bazel RequestMetadata in the
context and instead cause redundant unmarshalling of this proto.

This fix just installs those interceptors after the `copyHeaders`
interceptors, though this will now be done in the cli sidecar as well.

Related Issues:
https://github.com/buildbuddy-io/buildbuddy-internal/issues/5094