Add auth to context for index template (#9529)
This will allow experiment flags to be resolved correctly. The auth
information is included in the request, but wasn't being added to the
context, and so wasn't available to the experiment flag evaluator.

This PR also adds processing of the codesearch_enabled experiment, which
will control whether the "Enable Code Search" check box is shown on the
settings page. I'll add the flag definition separately to
buildbuddy-internal.