Codesearch: match filenames without explicit filter (#9581)
This change makes codesearch return matches from bare queries on
filenames, e.g. The query `buildfix` now returns the file `buildfix.sh`.

This PR is surprisingly large - this is because scoring needed a more
structured concept of "or" vs. "and" in order to score properly. The
current general factoring is maybe slight overkill for the specific case
we see in queries: matching bare queries to content OR filename, AND
matching filename filters to filenames... but I prefer this to a more
hardcoded approach, and it will make it easier to support boolean query
syntax if we ever decide to do that.