[oci resolver] read manifests from cache (disabled by default) (#9561)
This change adds the `read_manifests_from_cache` flag and will fetch
image and index manifests from the cache when enabled.

In order to avoid pathological behavior wherein `Layer.DiffID()` fetches
the entire uncompressed layer in order to compute a checksum, I also
implement `layerFromDigest` in this change. A future change will cache
the layer on calls to `Compressed()`.

---------

Co-authored-by: Brandon Duffany <brandon@buildbuddy.io>