Upgrade local grafana version (#9602)
This allows us to easily pin to a specific version of the grafana
clickhouse plugin, so we can work around
https://github.com/grafana/clickhouse-datasource/issues/1256.

I checked all the graphs, and they seem to work the same as the deployed
version. Unfortunately, the diff is huge because grafana ran some
migration when upgrading the version.