Flag-guard cluster-local reads/writes of workflow VM snapshots (#9639)
When I moved the dev GCP cache proxy to a deployment model instead of a
statefulset model, I broke snapshot sharing in dev. This is because
there's now no guarantee that an entire snapshot will be present in the
cache proxy that serves the snapshot read request. This PR adds a flag
that control whether or not to read/write snapshots from/to the backing
cache as well, so that we can do this in dev GCP, but not do it in SJC
where we anticipate continuing to run a distributed cache ring in the
Cache Proxies.