Codesearch: add DropNamespace operation on codesearch service (#9616)
Expose a DropNamespace operation, which can be used if the index format
is changed and reindexing is needed.

Operationally, this enables a relatively easy (although immature)
pattern when needing to force a reindex: manually call `DropNamespace`
on every codesearch-enabled namespace, then let the next incremental
index automatically trigger a full re-index.

Note this could be improved significantly - codesearch results will be
empty between the `DropNamespace` call and the next reindex. We could
drop the namespace and reindex all the repos in them in the same commit,
which would avoid any downtime. This is a more complex interface change
though, because it requires a) knowing (or asking the client to list)
each repo in a namespace, and b) plumbing in access tokens for each repo
in the namespace.

It's also not clear we would even use the production server for full
reindexing in a larger scale production environment - in those cases we
may want to compute a new index offline and swap it in to avoid a large
indexing load on the server.