Move user/group-specific experiments to GetUser (#9640)
We shouldn't send user- or group- specific data or experiment flags in
the `FrontendConfig` - it requires a full page refresh to update these.
Instead, this PR moves user-specific experiment data to `GetUser`.

- Add a new message to hold experiment data to `GetUserResponse`, and
populate it in the `GetUser` call
- Plumb that to the frontend `User` struct
- Update the frontend to check the new codesearch experiment flag when
deciding whether to render the "Enable Codesearch" checkbox on the
settings page.
- Set the selected group id cookie whenever group id changes.
- Undo the auth wrapping for root app requests from #9529, it is no
longer needed.