[oci resolver] use Puller to avoid re-authing to fetch layers (#9631)
Calls to `remote.Head(...)`, `remote.Get(...)`, and `remote.Layer(...)`
will cause the OCI client to re-authenticate to the remote registry.
This change creates a single `Puller` and uses it to fetch manifests and
layers, reducing the `GET /v2/` (and subsequent authentication)
requests. Each call to `Resolve(...)` still has to authenticate once.