[oci resolver] synchronize fetching config file & constructing remote Layer; add concurrent layer fetch test (#9634)
`ociruntime` fetches layers concurrently. During the course of fetching
layers, it also calls `DiffID()` on each layer, which can lead to
multiple concurrent calls of `RawConfigFile` on the image.

Making sure an image only fetches the raw config file from the remote
registry once cuts down on these unnecessary GETs.

This change introduces a test that calls DiffID and Compressed on
multiple layers concurrently.

This change also synchronizes creation of a `remote.Layer` inside
`layerFromDigest`. I'm not sure this is necessary, but it's
straightforward and the only other place where an Image or Layer struct
could be modified concurrently.