Unmount child mounts before cleaning build root (#9642)
To avoid traversing into overlayfs directories and writing tombstone
files (which is unnecessary and incurs a bunch of disk IO), unmount any
nested mounts before cleaning up the build root directory. If any
unmount fails, also skip those directories.

The implementation is somewhat complex because we can no longer use
`os.RemoveAll` as it doesn't have a way to skip files. Eventually we
probably want to move these mounts somewhere else (outside of the build
root) and/or address the root cause of the mounts sticking around so
that this will no longer be necessary:
https://github.com/buildbuddy-io/buildbuddy-internal/issues/5145