[FC] store creation and update timestamps in VMMetadata
This will allow us to expire snapshots that were created or updated a long time ago.

I'm still not sure if this is necessary, but storing the timestamps makes it easier to do this in the future. Also, it's cheap and it allows us to see how long snapshots survive.