Allow migration cache to be configured using experiment flags (#9641)
This will allow us to turn the migration on/off faster. We will also be
able to migrate one group at a time.

The experiment flag values are objects with 4 fields:
- state (src-only, src-primary, dest-primary, dest-only)
- async dest writes
- double read percentage
- decompress read percentage

If the experiment provider isn't available, or the flag is missing, or
one of the fields in the object is missing, we default to the global
config.

See https://github.com/buildbuddy-io/buildbuddy-internal/pull/5141 for
the initial experiment flag configurations, which will disable the
migration even once the migration cache is configured through regular
flags.