CLI: Register commands in a different place from where the command list is instantiated (#9635)
The parser needs to have access to the full list of CLI commands, and
many packages use or may need to use the parser. To avoid dependency
cycles, we declare the list in one package and populate using the
`Register` function in a sub-package. We call `register.Register` as one
of our first calls upon beginning execution.