Add a library to handle sequence / slice operations seamlessly (#9668)
The `seq` library has three main motivations:

- Frequently, when doing complex processing on a slice, intermediary
slice allocations are made to store intermediary values, or filtered
versions of the slice. These allocations can be quite significant,
depending on the size of the original slice. Using sequences prevents
these allocations, as the sequence is necessarily only looped over once
and does not store prior values, nor does it need to make copies in
order to remove filtered elements, as those are just skipped when
iterating.

- Similarly, lazy evaluation when processing a slice could lead to
saving cpu time if it turns out that we only needed to process part of
the slice. The sequence functions here are lazy, only processing a given
element at time of access, so early returns or breaks short-circuit the
rest of the processing,

- Sequences (`iter.Seq[E]`) in go are cumbersome to use, and
traditionally require separate functions for handling slices as opposed
to sequences or additional verbosity that affects readability (like
`slices.Values`). The functions in the `seq` library do not have this
problem, as they make use of the `Sequenceable[E any]` type constraint
and the `Sequence[E]` function, allowing them to accept either slices or
sequences without issue.