Fix preserve-workspace behavior on case-insensitive filesystems (#9690)
Fixes an issue that we would not properly replace old workspace inputs
in the case where a file is renamed in a way that changes only the case
of the file.

The problem was that we were tracking the known inputs in the workspace
using exact string matching, so we would fail to unlink pre-existing
input paths which are equivalent, but not exactly equal, to a new input
path.

The solution in this PR is to introduce a utility struct `fspath.Key`
which is a representation of a file path that treats differently-cased
file paths as being equal. It's internally just a lowercase string, but
the wrapper type adds some safety by forcing callers to take case
sensitivity into account.