Imrovements to navigation and references in code browser (#9707)
This PR has a few tangentially related changes (sorry for not breaking
it up more):

- Sort test files after non-test files in references panel
- Fetch and display references for multiple matching decorations if the
decorations share the same span
- This is not the common case, but comes up when libs are embedded into
go tests - the definitions in the embedded lib get their own, separate
nodes. This means that to show *all* references, we need to follow *all*
definitions associated with a given anchor.
- When clicking on a usage/call, navigate immediately to the definition.
- Fall back to displaying references if the definition isn't found (e.g.
if you click on a method from an external lib). We could detect and do
better with external references in the future...