Make migration_cache.config more lenient with missing or invalid configs (#9695)
Currently when flagd is not configured, we use
`openfeature.NoopProvider`, which always returns the default value.
Since this is nil, `config` always returns an error.

When flagd is configured, but the configuration is invalid, it also
returns the default, so again `config` always returns an error.

Instead log errors when the config is invalid or when flagd returns an
error. Don't log error when the config is missing, because this could be
just because flagd isn't configured.