Code browser changes to use extended xref kythe data (#9718)
We now show "extends", "extended by", "overrides", and "overridden by"
in the references panel. Additionally, references to code that is
related via superclasses, interfaces, and subclasses are now included in
the references section. For go, at least, this seems to actually,
finally, catch all references that you'd expect when clicking on a
symbol.

Depends on #9715