CLI: Add "native" parser (#9734)
Adds (and makes use of) a parser that can handle the `--verbose` (and
`-v`), `--watch` (and `-w`) and `watcher_flags`.

This prevents the cli from grabbing arguments that look like these
options but are not. These options are defined as startup options, which
means they must appear before the command.