invocation_action_card: don't skip fetchActionResult
In #8114, we replaced fetchActionResult with streamExecution so that we
can view in-progress RBE actions using WaitExecution RPC call.

However, this breaks cache-only invocations when trying to navigate from
Cache request's mnemonic to the action page. In a cache-only build,
there will be no Execution and thus, fetchActionResult is the only way
user can see the action result.

Fix this by calling fetchActionResult together with streamExecution.