Don't parse auth headers twice between atime_updater and hit_tracker_client (#9740)
This is a little bit trickier than throwing it in an interceptor and
calling it a day because these headers are only used in the
atime_updater and the hit_tracker_client, and those are only called for
requests that are served locally, so it's most efficient to do this work
only when we know that's how the request will be processed. This is my
attempt at that. 😛

Related issues:
https://github.com/buildbuddy-io/buildbuddy-internal/issues/5094