CLI: Add "help" parser (#9737)
This makes the `-h` and `--help` options work. It required what I think
are minimally invasive changes to the parser code to allow for stuff
like `bb -h -- test`, as well as creating new option definitions to
allow for help options with noo explicit help command, but other than
that it is mostly able to do normal-ish bazel option parsing.